ANE-35 Bind CommunicationRequest to request
This commit adds changes that bind the CommunicationRequest object to
the request object. The middleware added performs resourceType
validation and bind the request.body to the request object. Upon
validation failure, an expection is thrown with the HTTP status code 400
and a custom error message.

I have chosen to take this middleware approach for validation so that
our request handlers can be kept clean and only focused use case logic.