Return OperationOutcome for POST CommunicationResource
This commit adds changes that update the http response message for the
POST CommunicationResource functionality. This commit also adds a
"Location" header with the value being a reference to the
CommunicationResource being processed. This reference will, in the
future, allow alert reporters for query the API using these generated
IDs.