Snapshot commit
This commit contains a lot changes that were initially not planned to be
in it. Due to the circumstance in the project, we have decided to make
this big commit and have one member(Rowan Pillay) of the team look at all of the
codebase in one go.

NOTE: This situation is not ideal and should not happen
again