ane-61 Add interface for sms channels
This commit contains changes include a new sms channel interface.
This is necessary to have an interface per channel(sms, WhatsApp, mobile app, etc)
to limit the amount of work done by the channels. We previously planned to let all
channels implement the `IChannel` interface, this was a bad design as it would have
forced every implmentation to know how to parse a `CommunicationRequest` resource.

This new interface expects only data necessary for sending an sms. All the
`CommunicationRequest` resource parsing and `Communication` resource creation
will be handled by the `channels` module itself.